testing/web-platform/tests/workers/WorkerGlobalScope_addEventListener.htm

<!DOCTYPE html>
<title> WorkerGlobalScope API: addEventListener() </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  var worker = new Worker('./support/EventListener.js');
  worker.onmessage = t.step_func_done(function(e) {
    assert_equals(e.data, "ping");
  });
  worker.postMessage("ping");
});
</script>
